requirements.txt
django >= 3.2, < 3.3
djangorestframework >= 3.12, < 3.13
